Реализация сервисов

+++ HikikomoriWEB.Services/RepositoryServices/RateContentService.cs
@@ -54,6 +54,14 @@ namespace HikikomoriWEB.Services.RepositoryServices
             try
             {
                 var response = new ResponseRepository<RateContent>();
+
+                if(_repository.GetOnId(ContentId) == null)
+                {
+                    response.Description = $"RateContent.Method [DeleteContent] : Элемент таблицы с таким ID не найден";
+                    response.StatusCode = StatusCode.NotFound;
+                    return response;
+                }
+
                 await _repository.DeleteContent(ContentId);
                 response.StatusCode = StatusCode.OK;
                 return response;
